feat: move FundingDialog
